# improved EntryNotFoundError