--- a/entry_not_found_error.go
+++ b/entry_not_found_error.go
@@ -4,14 +4,14 @@ import "fmt"
 
 // EntryNotFoundError is an error type struct which is returned when entry was not found for provided key
 type EntryNotFoundError struct {
-	message string
+	key string
 }
 
 func notFound(key string) error {
-	return &EntryNotFoundError{fmt.Sprintf("Entry %q not found", key)}
+	return &EntryNotFoundError{key}
 }
 
 // Error returned when entry does not exist.
 func (e EntryNotFoundError) Error() string {
-	return e.message
+	return fmt.Sprintf("Entry %q not found", e.key)
 }