- [x] error message will be generated on demand
- [x] added bench readFromCacheNonExistentKeys

before:
```
BenchmarkReadFromCacheNonExistentKeys/1-shards-4         	 5000000	       275 ns/op	      71 B/op	       3 allocs/op
BenchmarkReadFromCacheNonExistentKeys/512-shards-4       	 5000000	       291 ns/op	      71 B/op	       3 allocs/op
BenchmarkReadFromCacheNonExistentKeys/1024-shards-4      	 5000000	       269 ns/op	      71 B/op	       3 allocs/op
BenchmarkReadFromCacheNonExistentKeys/8192-shards-4      	 5000000	       277 ns/op	      71 B/op	       3 allocs/op
```


after:
```
BenchmarkReadFromCacheNonExistentKeys/1-shards-4         	10000000	       197 ns/op	      23 B/op	       1 allocs/op
BenchmarkReadFromCacheNonExistentKeys/512-shards-4       	10000000	       187 ns/op	      23 B/op	       1 allocs/op
BenchmarkReadFromCacheNonExistentKeys/1024-shards-4      	10000000	       159 ns/op	      23 B/op	       1 allocs/op
BenchmarkReadFromCacheNonExistentKeys/8192-shards-4      	10000000	       177 ns/op	      23 B/op	       1 allocs/op
```